Move config related functions to minimalin.c
